+++ src/mvvm/awtk/binding_context_awtk.h
@@ -32,11 +32,13 @@ BEGIN_C_DECLS
  * 创建binding_context对象。
  *
  * @param {widget_t*} widget 控件对象。
+ * @param {binding_context_t*} parent_ctx 父ctx。
  * @param {navigator_request_t*} req 请求参数。
  *
  * @return {binding_context_t*} 返回binding_context对象。
  */
-binding_context_t* binding_context_awtk_create(widget_t* widget, navigator_request_t* req);
+binding_context_t* binding_context_awtk_create(widget_t* widget, binding_context_t* parent_ctx,
+                                               navigator_request_t* req);
 
 /*public for test*/
 ret_t binding_context_bind_for_window(widget_t* widget, navigator_request_t* req);