+++ src/mvvm/jerryscript/jerry_script_helper.c
@@ -314,4 +314,3 @@ ret_t jerry_script_register_builtins(void) {
 
   return RET_OK;
 }
-